Helpers are services, consolidate namespaces.
