+++ config/packages/twig.yaml
@@ -6,7 +6,7 @@ twig:
         default_locale: "%kernel.default_locale%"
         supported_locales: "%supported_locales%"
         cards_data: '@App\Services\CardsData'
-        deck_validation_helper: '@App\Helper\DeckValidationHelper'
+        deck_validation_helper: '@App\Services\DeckValidationHelper'
         google_analytics_tracking_code: "%google_analytics_tracking_code%"
         allow_search_engine_tracking: "%allow_search_engine_tracking%"
     form_themes:
