+++ src/Services/DeckManager.php
@@ -11,8 +11,8 @@ use App\Entity\Deckslot;
 use App\Entity\FactionInterface;
 use App\Entity\PackInterface;
 use App\Entity\UserInterface;
-use App\Helper\AgendaHelper;
-use App\Helper\DeckValidationHelper;
+use App\Services\AgendaHelper;
+use App\Services\DeckValidationHelper;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\OptimisticLockException;
